total to double

+++ lib/features/cart/domain/entites/cart_entity.dart
@@ -3,12 +3,12 @@ import 'package:collection/collection.dart';
 import 'cart_entity/sub_cart_entity.dart';
 
 class CartEntity {
-  final int? totalPrice;
+  final double? totalPrice;
   final List<SubCartEntity>? data;
 
   CartEntity({this.totalPrice, this.data});
 
-  CartEntity copyWith({List<SubCartEntity>? data, int? totalPrice}) {
+  CartEntity copyWith({List<SubCartEntity>? data, double? totalPrice}) {
     return CartEntity(
       data: data ?? this.data,
       totalPrice: totalPrice ?? this.totalPrice,
